refactor: rename test utils to lowercase
