+++ tests/cogs/announce/test_cog.py
@@ -10,13 +10,13 @@ from discord.ext import commands
 # Own modules
 import koalabot
 from koala.cogs import announce
-from tests.tests_utils import LastCtxCog
+from tests.tests_utils.last_ctx_cog import LastCtxCog
 
 from tests.log import logger
 
 @pytest.fixture(autouse=True)
 def utils_cog(bot: discord.ext.commands.Bot):
-    utils_cog = LastCtxCog.LastCtxCog(bot)
+    utils_cog = LastCtxCog(bot)
     bot.add_cog(utils_cog)
     dpytest.configure(bot)
     logger.info("Tests starting")
