+++ tests/tests_utils/__init__.py
@@ -1,1 +1,1 @@
-from . import fake_load_all_cogs, LastCtxCog, TestUtils
+from . import fake_load_all_cogs, last_ctx_cog, utils
