+++ tests/tests_utils/fake_load_all_cogs/__init__.py
@@ -1,1 +1,1 @@
-from . import Greetings
+from . import greetings_cog
